==> 2_Sampling_Outer_Inner_CV_Class_Labels.ipynb <==
#!/usr/bin/env python
# coding: utf-8

if 'config' not in locals():
    config = 'configs.config_ILLAE_local'
#     config = 'configs.config_WomenInEngineering_local'
#     config = 'configs.config_Digits_local'    

### initialization ###

import importlib
print('config file is set to {}'.format(config))

imp=importlib.import_module('_imports')
importlib.reload(imp)

util=importlib.import_module('_utils')
importlib.reload(util)

c=importlib.import_module(config)
importlib.reload(c)

get_ipython().run_line_magic('matplotlib', 'inline')

# param_setup(sys.argv[1:], c)
util.param_setup_ipython(globals(), c)

### reading the dataset file ###

dfs=dict()
dfs['input']=imp.pd.read_csv('{}data/{}_1_data_filtered.csv'.format(c.directory['save'], c.project_name))

imp.pprint(dfs['input'].shape)
util.display_df(dfs['input'].head())

### creating label dataframe + saving ###

# dfs['label'] = imp.pd.DataFrame()
# dfs['input'][c.id_name] = dfs['input'][c.id_name]
if c.label is not None:
    if c.label['type'] == 'boolean':
        dfs['input']['label'] = dfs['input'][c.label['col']] > 0
        dfs['input']['label'] = dfs['input']['label'].astype(str)

    dfs['input'][[c.id_name, 'label']].to_csv('{}data/{}_2_data_labels.csv'.format(c.directory['save'], c.project_name), index=False)    
    util.display_df(dfs['input'].head(5))

### calculating stratified outer + inner folds and saving ###

dfs['input']=imp.pd.read_csv('{}data/{}_2_data_labels.csv'.format(c.directory['save'], c.project_name))
class_labels = set(dfs['input']['label'])

samples=min(dfs['input']['label'].value_counts())
print('Number of samples per each class = {}'.format(samples))
util.mkdir(c.directory['save']+'cv_ids/')

ids_list=list()
labels_list=list()

for label in class_labels:
    ids=list(imp.np.random.choice(dfs['input'].loc[dfs['input']['label']==label, c.id_name], samples, replace=False))
    labels=[label]*samples
    ids_list.extend(ids)
    labels_list.extend(labels)
    
skf_outer = imp.StratifiedKFold(n_splits=c.outer, random_state=c.seed)
for outer_index, (train_index, test_index) in enumerate(skf_outer.split(ids_list, labels_list)):
#     print (test_index)

    outer_ids = [ids_list[id] for id in test_index]
    outer_labels = [labels_list[id] for id in test_index]
    inner_ids = [ids_list[id] for id in train_index]
    inner_labels = [labels_list[id] for id in train_index]
    
#     util.display_df(imp.pd.DataFrame([outer_ids, outer_labels]).T)
    df=imp.pd.DataFrame(dict(id=outer_ids))
    df.to_csv('{}cv_ids/outer_{}.csv'.format(c.directory['save'], outer_index), index=False)

    skf_inner = imp.StratifiedKFold(n_splits=c.inner, random_state=c.seed)
    for inner_index, (train_index, test_index) in enumerate(skf_inner.split(inner_ids, inner_labels)):
        inner_ids_final = [inner_ids[id] for id in test_index]
        inner_labels_final = [inner_labels[id] for id in test_index]

#         util.display_df(imp.pd.DataFrame([inner_ids_final, inner_labels_final]).T)
        df=imp.pd.DataFrame(dict(id=inner_ids_final))
        df.to_csv('{}cv_ids/inner_{}_{}.csv'.format(c.directory['save'], outer_index, inner_index), index=False)
