==> 3_Static_Feature_Creation.ipynb <==
#!/usr/bin/env python
# coding: utf-8

if 'config' not in locals():
#     config = 'configs.config_WomenInEngineering_local'
    config = 'configs.config_ILLAE_local'    
#     config = 'configs.config_Juul_local'        

### initialization ###

import importlib
print('config file is set to {}'.format(config))

imp=importlib.import_module('_imports')
importlib.reload(imp)

util=importlib.import_module('_utils')
importlib.reload(util)

c=importlib.import_module(config)
importlib.reload(c)

get_ipython().run_line_magic('matplotlib', 'inline')

# param_setup(sys.argv[1:], c)
util.param_setup_ipython(globals(), c)

### functions for feature creation ###

def hashtags_counter(hashtags):
    if not isinstance(hashtags, str):
#         raise Exception('{} is not string. It is {}'.format(hashtags, type(hashtags)))
        print('{} is not string. It is {}'.format(hashtags, type(hashtags)))
        return [0, 0]
    hashtags_list=hashtags.split(',')
    sum_len = 0
    for h in hashtags_list:
        sum_len += len(h.strip())
    return [len(hashtags_list), sum_len/len(hashtags_list)]

def concat_cols(row):
    return ' '.join(map(str, row))

def concat_cols_df(df):
    if isinstance(df, imp.pd.DataFrame):
        if len(df.columns) == 1:
            return df
        return df.apply(concat_cols, axis=1)
    elif isinstance(df, imp.pd.Series):
        return df.apply(concat_cols)
    raise util.Failed('Input was not either dataframe or series!')

def preprocess_text(text, replacements):
#     replacements=list(replacements)
#     print(replacements)
    for replacement in replacements:
        try:
            text = imp.re.sub(*replacement, text)
        except TypeError as e:
            print('{} + text {}'.format(replacement, text))
            raise e
#     return tokenizer_tag(text)
    return text

def preprocess_text_series(series, replacements):
    return series.apply(preprocess_text, args=(replacements, ))

def calc_age_of_acct(acct_date):
    first_date=imp.date(2005, 1, 1)
    delta = acct_date.date() - first_date
    return delta.days

def calc_age_of_acct_series(series):
    return series.apply(calc_age_of_acct)

def char_len(text):
    return len(text)
    
def char_len_series(series):
    return series.apply(char_len)

def check_words(txt, words):
    txt=' '+txt.lower()+' '
    for word in words:
        word= ' ' + word + ' '
        if word in txt:
            return 1
    return 0

def check_words_series(words, series):
    return series.apply(check_words, args=(words,))

def anew(txt, emo_dict):
    words = imp.re.findall('[a-z]+', txt.lower())
    result={'valence':0,
            'dominance':0,
            'arousal':0
           }
    for key, emo in emo_dict.items():
        for word in words:
            if word in emo:
                result[key] += emo[word]
    return result

def anew_series(emo_dict, series):
    return series.apply(anew, args=(emo_dict,))

### preping the input file ###

util.mkdir('{}data/features/'.format(c.directory['save']))

dfs=dict()
dfs['input']=imp.pd.read_csv('{}data/{}_1_data_filtered.csv'.format(c.directory['save'],
                                                                    c.project_name))

### creating static (meta) features and saving ###

for name, feature in c.features['static'].items():
    if feature['type']=='meta':
        cols=list([c.id_name])
        cols.extend(feature['cols'].keys())
        dfs[name]=dfs['input'][cols].copy()
        for col, col_type in feature['cols'].items():
            print('name = {}, type = {}'.format(col, col_type))

            if col_type == 'boolean':
                dfs[name][col] = (imp.pd.notna(dfs[name][col])).astype(float)
            elif col_type == 'age_of_account':
                dfs[name][col] = util.parallelize_df(dfs[name][col], util.str_date_to_date_df, c)
                dfs[name][col] = util.parallelize_df(dfs[name][col], calc_age_of_acct_series, c)
            elif col_type == 'length':
                dfs[name][col] = util.parallelize_df(dfs[name][col], char_len_series, c)
            elif col_type == 'lookup':
                dfs[name][col]=util.parallelize_df(dfs[name][col], check_words_series, c, 
                                                   params=(feature['words'], ))
            elif col_type == 'anew':
                ANEW = imp.pd.read_csv('./ANEW.csv')
                valence=dict(zip(ANEW['Word'], ANEW['valence']))
                dominance=dict(zip(ANEW['Word'], ANEW['dominance']))
                arousal=dict(zip(ANEW['Word'], ANEW['arousal']))

                emo_dict={'valence':valence,
                          'dominance':dominance,
                          'arousal':arousal
                         }
                results=util.parallelize_df(dfs[name][col], anew_series, c, params=(emo_dict,))
                dfs[name] = imp.pd.concat([dfs[name],
                                           imp.pd.DataFrame(list(results))
                                          ]
                                           , axis=1)
                dfs[name].drop([col], axis=1, inplace=True)
            elif col_type == 'hashtag':
                result=dfs[name][col].apply(hashtags_counter)
                result=list(zip(*result))
                dfs[name] = imp.pd.concat([dfs[name], 
                                           imp.pd.DataFrame(zip(*result), columns=['hashtags_count', 'hashtags_sum'])
                                          ]
                                          , axis=1) 
                dfs[name].drop([col], axis=1, inplace=True)

        dfs[name].to_csv('{}data/features/{}_feature_{}.csv'.format(c.directory['save'], 
                                                                   c.project_name,
                                                                   name
                                                                  ), index=False)
        imp.pprint('name of the dataframe = {}'.format(name))
        util.display_df(dfs[name].head())

### creating static (non-meta) features and saving ###

for  name, feature in c.features['static'].items():
    if feature['type']=='liwc':

        dfs[name]=imp.pd.read_csv('{}data/raw/{}'.format(c.directory['master'], 
                                                            feature['filename']))
        imp.pprint(dfs[name].shape)
        util.display_df(dfs[name].head())
    #     print(list(dfs[name]))

        index_id=list(dfs[name].iloc[0,:]).index(c.id_name)
        index_wc=list(dfs[name]).index('WC')
        dfs[name]=imp.pd.concat([dfs[name].iloc[1:, index_id], 
                                   dfs[name].iloc[1:, index_wc:]], 
                                  axis=1)
        col_names=list(dfs[name])
        col_names[0]=c.id_name
        dfs[name].columns=col_names
        dfs[name].to_csv('{}data/features/{}_feature_{}.csv'.format(c.directory['save'], 
                                                                    c.project_name,
                                                                    feature['type']), 
                         index=False)
        imp.pprint(dfs[name].shape)
        util.display_df(dfs[name].head())

### preparing data for dynamic features and saving ###

dynamics=dict()

for name, feature in c.features['dynamic'].items():
    print('{} - {}'.format(feature['type'], name))
    cols=list([c.id_name])
    cols.extend(feature['cols'])
    df_dynamic=dfs['input'][cols].copy()

#   the nodes column should be have comma as separator
    if feature['type'] not in dynamics:
        dynamics[feature['type']]=importlib.import_module('dynamic_features.{}'.format(feature['type']))
        importlib.reload(dynamics[feature['type']])
    df_dynamic = dynamics[feature['type']].prep_features(df_dynamic, feature, c)    
    df_dynamic.to_csv('{}data/features/{}_feature_{}({}).csv'.format(c.directory['save'], 
                                                                     c.project_name,
                                                                     feature['type'],
                                                                     name), 
                      index=False)
