==> 4_Experiments_Setup.ipynb <==
#!/usr/bin/env python
# coding: utf-8

if 'config' not in locals():
    config = 'configs.config_ILLAE_local'
#     config = 'configs.config_WomenInEngineering_local'
#     config = 'configs.config_Digits_local'    

### initialization ###

import importlib
print('config file is set to {}'.format(config))

imp=importlib.import_module('_imports')
importlib.reload(imp)

util=importlib.import_module('_utils')
importlib.reload(util)

c=importlib.import_module(config)
importlib.reload(c)

get_ipython().run_line_magic('matplotlib', 'inline')

# param_setup(sys.argv[1:], c)
util.param_setup_ipython(globals(), c)

### creating parameters grid for the algorithm hyperparameters ###

experiments_cv=[]

if c.algorithm['name'] == 'svm':
    for kernel in c.algorithm['kernels']:
#         params_svm=imp.copy.deepcopy(c.algorithm['params'])
        params=c.algorithm['params']
        params_svm=dict()
        params_svm['cost'] = imp.copy.deepcopy(params['cost'])    

        if kernel != 'linear':
            params_svm['gamma'] = imp.copy.deepcopy(params['gamma'])
        if kernel == 'poly':
            params_svm['degree'] = imp.copy.deepcopy(params['degree'])
        if kernel == 'poly' or kernel == 'sigmoid':        
            params_svm['coef0'] = imp.copy.deepcopy(params['coef0'])

        keys, values = zip(*params_svm.items())
        experiments = [dict(zip(keys, v)) for v in imp.itertools.product(*values)]
        print('number of experiments with {} kernel = {}'.format(kernel, len(experiments)))
        result_exps=[]

        for experiment in experiments:
            experiment['kernel']=kernel

            for cv in range(c.inner):
                experiment['cv']=cv
                experiments_cv.append(imp.copy.deepcopy(experiment))

experiments_cv_df=imp.pd.DataFrame(experiments_cv)
experiments_cv_df.fillna(value=1, inplace=True)

print(experiments_cv_df.shape)
util.display_df(experiments_cv_df.head(11))

### adding dynamic features parameters to the parameter search ###

for feature in c.select_features['dynamic']:
    for param_name, param_list in c.features['dynamic'][feature]['params'].items():
        param_name='{}_{}'.format(feature, param_name)
        print(param_name, param_list)
        
        experiment_cvs_updated=[]
        for param_val in param_list:
            for experiment in experiments_cv:
                experiment_copy=imp.copy.deepcopy(experiment)
                experiment_copy[param_name]=param_val
                experiment_cvs_updated.append(experiment_copy)
        experiments_cv=experiment_cvs_updated        

experiments_cv_df=imp.pd.DataFrame(experiments_cv)
experiments_cv_df.fillna(value=1, inplace=True)

print(experiments_cv_df.shape)
util.display_df(experiments_cv_df.head(15))

### saving parameters grid ###

util.mkdir('{}experiments/{}/'.format(c.directory['save'], util.experiment_name(c)))
experiments_cv_df.to_csv('{}experiments/{}/1_parameters.csv'.format(c.directory['save'],util.experiment_name(c)),
                                                                  index=False)

### creating parameters grid for dynamic feature parameters ###

experiments_cv=[]
experiment=dict()

for cv in range(c.inner):
    experiment['cv']=cv
    experiments_cv.append(imp.copy.deepcopy(experiment))

experiments_dynamics=[]
for feature in c.select_features['dynamic']:
    experiments_cv_copy=imp.copy.deepcopy(experiments_cv)
    for param_name, param_list in c.features['dynamic'][feature]['params'].items():
        param_name='{}_{}'.format(feature, param_name)
        print(param_name, param_list)
        
        experiment_cvs_updated=[]
        for param_val in param_list:
            for experiment in experiments_cv_copy:
                experiment_copy=imp.copy.deepcopy(experiment)
                experiment_copy[param_name]=param_val
                experiment_copy['name']=feature
                experiment_copy['type']=c.features['dynamic'][feature]['type']                
                experiment_cvs_updated.append(experiment_copy)
        experiments_cv_copy=experiment_cvs_updated
    experiments_dynamics.extend(experiments_cv_copy)

experiments_cv_df=imp.pd.DataFrame(experiments_dynamics)
experiments_cv_df.fillna(value='na', inplace=True)

print(experiments_cv_df.shape)
util.display_df(experiments_cv_df.head(11))

### saving dynamic parameters grid ###

experiments_cv_df.to_csv('{}experiments/{}/1_parameters_dynamic_features.csv'.format(c.directory['save'],
                                                                                     util.experiment_name(c)),
                         index=False)
