==> 7_Optimal_Model.ipynb <==
#!/usr/bin/env python
# coding: utf-8

if 'config' not in locals():
    config = 'configs.config_ILLAE_local'
#     config = 'configs.config_WomenInEngineering_local'
#     config = 'configs.config_Digits_local'        

### initialization ###

import importlib
print('config file is set to {}'.format(config))

imp=importlib.import_module('_imports')
importlib.reload(imp)

util=importlib.import_module('_utils')
importlib.reload(util)

c=importlib.import_module(config)
importlib.reload(c)

get_ipython().run_line_magic('matplotlib', 'inline')

# param_setup(sys.argv[1:], c)
util.param_setup_ipython(globals(), c)

### reading experiment results from pickled files ###

dbfile = open('{}tmp/{}_db_result_cvs_1_from_{}.pickle'.format(c.directory['save'],
                                                               util.experiment_name(c),
                                                               c.index_step
                                                              ), 'rb') 
result_cvs=imp.pickle.load(dbfile)
dbfile.close()

for i in range(2, c.index_step+1):
    dbfile = open('{}tmp/{}_db_result_cvs_{}_from_{}.pickle'.format(c.directory['save'],
                                                           util.experiment_name(c),
                                                           i,
                                                           c.index_step
                                                              ), 'rb') 
    result_cvs.extend(imp.pickle.load(dbfile))
    dbfile.close()

imp.pprint('number of experiments = {}'.format(len(result_cvs)))

### functions for metric evaluation ###

def metric_eval(pred_class, true_class, class_labels):
    accuracy=0
    for pred, true in zip(pred_class, true_class):
        if pred == true:
            accuracy += 1
    accuracy /= len(pred_class)
    
    F1_score=dict()
    for c in class_labels:
        TP = 0
        FP = 0
        FN = 0
        for pred, true in zip(pred_class, true_class):
            if pred == true and pred == c:
                TP += 1
            elif pred != true and pred == c:
                FP += 1
            elif pred != true and true == c:
                FN += 1
        if TP == 0:
            F1_score[c] = 0
            continue

        precision = TP/(TP+FP)
        recall = TP/(TP+FN)
        F1_score[c] = 2 / (1/precision + 1/recall)
        
    return dict(
        accuracy=accuracy, 
        F1_score=F1_score
    )
    
def create_list_of_results(df_to_show):
    df_to_show= imp.pd.concat([imp.pd.DataFrame([e['eval']['F1_score'] for e in df_to_show]), 
                               imp.pd.DataFrame(dict(accuracy=[e['eval']['accuracy'] for e in df_to_show])), 
                               imp.pd.DataFrame([e['params'] for e in df_to_show])
                              ], axis=1)
    df_to_show.columns=[str(label) for label in list(df_to_show)]
    return df_to_show

### calculating evaluation metrics for each experiment ###

dfs=dict()
dfs['labels']=imp.pd.read_csv('{}data/{}_2_data_labels.csv'.format(c.directory['save'], c.project_name))
# class_labels = [str(label) for label in set(dfs['labels']['label'])]
class_labels = set(dfs['labels']['label'])

keys=list(result_cvs[0]['experiment'].keys())
del keys[keys.index('cv')]
result_cvs=sorted(result_cvs,
                  key=lambda i:[i['experiment'][col] for col in keys]
                 )    
experiment_results=[]

for i in range(0, len(result_cvs), c.inner):
    pred_list=[]
    true_list=[]

    for j in range(c.inner):
        pred_list.extend(result_cvs[i+j]['pred'])
        true_list.extend(result_cvs[i+j]['true'])

    params=result_cvs[i+j]['experiment']
    params.pop('cv', None)
    experiment_results.append({
        'params':params
        , 'eval':metric_eval(pred_list, true_list, class_labels)
    })
 
params_search_results=create_list_of_results(experiment_results)

params_search_results.to_csv('{}experiments/{}/2_parameters_cv_results.csv'.format(c.directory['save'],
                                                                                   util.experiment_name(c)),
                             index=False)
util.display_df(params_search_results.head(100))

### finding the optimal paramters and saving ###

max_results=list()

max_result=imp.copy.deepcopy(params_search_results.sort_values('accuracy', ascending=False).iloc[0])
max_result=max_result.append(imp.pd.Series(['accuracy'], index=['metric']))
max_results.append(max_result)

for label in list(map(str, class_labels)):
#     print(label)
    max_result=imp.copy.deepcopy(params_search_results.sort_values(label, ascending=False).iloc[0])
    max_result=max_result.append(imp.pd.Series([label], index=['metric']))
    max_results.append(max_result)

max_results_df=imp.pd.DataFrame(max_results)
max_results_list=[]
for max_result in list(map(dict, max_results)):
    max_result['cv']='all'
    max_results_list.append(max_result)
index=['accuracy']
index.extend(list(map(str, class_labels)))
max_results_df.index=index
max_results_df.to_csv('{}experiments/{}/3_optimal_parameters.csv'.format(c.directory['save'],
                                                                         util.experiment_name(c)),
                      index=True)    
util.display_df(max_results_df)

### reading class labels and concatenating features and original data ###

dfs=dict()

dfs['labels'] = imp.pd.read_csv('{}data/{}_2_data_labels.csv'.format(c.directory['save'], c.project_name))
dfs['features']=util.construct_static_features(c)
if len(dfs['features']) == 0:
    dfs['features'] = imp.pd.DataFrame(dfs['labels'][c.id_name])
    
imp.pprint('labels = {}'.format(dfs['labels'].shape))
util.display_df(dfs['labels'].head())

imp.pprint('featuers = {}'.format(dfs['features'].shape))
util.display_df(dfs['features'].head())

### reading cv ids for cv folds ###

outer_ids = imp.pd.read_csv('{}{}'.format(c.directory['save'], 'cv_ids/outer_0.csv'))
outer_ids.rename(columns={'id': c.id_name}, inplace=True)

inner_ids = list()
for i in range(c.inner):
    inner_ids.append(imp.pd.read_csv('{}cv_ids/inner_0_{}.csv'.format(c.directory['save'], i)))

imp.pprint('Number of cv = {}'.format(len(inner_ids)))

### function for training and testing the optimal parameters ###

def optimal_model_training_testing(optimal_params, df_features, df_labels, df_inputs, dynamics):
    training = imp.copy.deepcopy(inner_ids)
    training_ids=imp.pd.DataFrame()
    for item in training:
        training_ids=imp.pd.concat([training_ids, item], ignore_index=True, axis=0)
    training_ids.rename(columns={'id':c.id_name}, inplace=True)
    testing_ids = imp.copy.deepcopy(outer_ids)
        
    for feature in c.select_features['dynamic']:
        dynamic=dynamics[c.features['dynamic'][feature]['type']]
        print(c.features['dynamic'][feature]['type'])
        df_input=df_inputs[feature]
        df_training=imp.pd.merge(df_input, training_ids, on=c.id_name, how='right')
        df_testing=imp.pd.merge(df_input, testing_ids, on=c.id_name, how='right')        
        optimal_params_temp=imp.copy.deepcopy(optimal_params)
        optimal_params_temp['name']=feature
        optimal_params_temp['type']=c.features['dynamic'][feature]['type']
        
        df_dynamic=dynamic.construct_features(c, optimal_params_temp, feature,
                                              df_training, df_testing,
                                              training_ids, testing_ids)
#         df_dynamic=dynamic.retrieve_features(c, optimal_params_temp, feature)
#         util.display(df_dynamic.head())

        df_features=imp.pd.merge(df_features, df_dynamic, on=c.id_name, how='right')

    result_training = util.data_prepropressing(training_ids, 
                                               df_features,
                                               df_labels,
                                               c
                                              )

    result_testing = util.data_prepropressing(testing_ids, 
                                         df_features, 
                                         df_labels,
                                         c,
                                         result_training['imputer'], 
                                         result_training['scaler']
                                        )    

    classifier = imp.SVC(C=optimal_params['cost'] if 'cost' in optimal_params else 1,
                         kernel=optimal_params['kernel'] if 'kernel' in optimal_params else 'linear',
                         gamma=optimal_params['gamma'] if 'gamma' in optimal_params else 1,
                         coef0=optimal_params['coef0'] if 'coef0' in optimal_params else 1,
                         degree=optimal_params['degree'] if 'degree' in optimal_params else 1,
                         max_iter=c.max_iter_optimization,
                         decision_function_shape='ovo', 
                         cache_size=c.cache_size,
                         random_state=c.seed
                        )
    
    classifier.fit(result_training['features'], list(result_training['labels']['label']))  
    array=imp.np.array(result_testing['features'])
    if len(result_testing['features'])==1:
        array=array.reshape(1, -1)

    predictions=classifier.predict(array)
    pred_list=list(predictions)
    true_list=list(result_testing['labels']['label'])
    test_result=dict(
        eval=metric_eval(pred_list, true_list, class_labels), 
        params=optimal_params
    )
    test_predictions = imp.pd.concat([
        outer_ids.astype(str),
        imp.pd.DataFrame(pred_list, columns=['pred']), 
        imp.pd.DataFrame(true_list, columns=['true'])], 
        axis=1)
    
    return dict(result=test_result, predictions=test_predictions)

### loading dynamic feature modules ###:

dynamics=dict()
df_inputs=dict()
for feature in c.select_features['dynamic']:
    dynamics[c.features['dynamic'][feature]['type']]=importlib.import_module('dynamic_features.{}'.format(c.features['dynamic'][feature]['type']))
    importlib.reload(dynamics[c.features['dynamic'][feature]['type']])
    df_inputs[feature]=imp.pd.read_csv('{}data/features/{}_feature_{}({}).csv'.format(c.directory['save'],
                                                                                         c.project_name,
                                                                                         c.features['dynamic'][feature]['type'],
                                                                                         feature))
    df_inputs[feature]=dynamics[c.features['dynamic'][feature]['type']].prep_input(df_inputs[feature], c)

### Debugging 1 - optimal params ###

# type(max_results[0])
# dict(max_results[0])
# max_result_test=max_results_list[0]
# optimal_params_temp=imp.copy.deepcopy(optimal_params)
# optimal_params_temp['name']=feature
# optimal_params_temp['type']=c.features['dynamic'][feature]['type']
# max_result_test

### Debugging 2 - running the for loop ###

# params=dict(
#     df_features=dfs['features'], 
#     df_labels=dfs['labels'], 
#     df_inputs=df_inputs,
#     dynamics=dynamics
# )

# optimal_model_training_testing(max_result_test, **params)

### running optimal models in parallel ###

optimal_results=imp.Parallel(n_jobs=c.num_cores
                    , prefer="processes"
                    , verbose=3
                   ) (imp.delayed(optimal_model_training_testing)(max_result, 
                                                                  dfs['features'], 
                                                                  dfs['labels'],
                                                                  df_inputs,
                                                                  dynamics
                                                                 ) 
                      for max_result in max_results_list)

### saving the optimal models results ###

optimal_results_df=create_list_of_results([optimal_result['result'] for optimal_result in optimal_results])
index=['accuracy']
index.extend(list(map(str, class_labels)))
max_results_df.index=index
optimal_results_df.index=index
optimal_results_df.to_csv('{}experiments/{}/4_optimal_models_results.csv'.format(c.directory['save'],
                                                                                 util.experiment_name(c)),
                      index=False)
util.display_df(optimal_results_df)

### saving the results of optimal models predictions ###

util.mkdir('{}experiments/{}/true_pred_results/'.format(c.directory['save'],
                                                        util.experiment_name(c)))
for optimal_result in optimal_results:
    optimal_result['predictions'].to_csv('{}experiments/{}/true_pred_results/predictions_{}.csv'.format(c.directory['save'],
                                                                                                        util.experiment_name(c),
                                                                                                        optimal_result['result']['params']['metric']
                                                                                                       ),
                                         index=False)
